Add xnnn basic opcodes

+++ src/cpu.rs
@@ -3,14 +3,14 @@ use crate::instructions::get_opcode_func;
 #[derive(Copy, Clone)]
 pub struct Cpu {
     pub registers: [u8; Cpu::REGISTERS_SIZE],
-    pub ram: [u8; Cpu::RAM_SIZE],
+    pub ram: [u8; Cpu::RAM_SIZE as usize],
     pub stack: [u16; Cpu::STACK_SIZE],    
-    pub pc: usize,
+    pub pc: u16,
     pub sp: u8,
 }
 
 impl Cpu {
-    pub const RAM_SIZE: usize = 0x1000;
+    pub const RAM_SIZE: u16 = 0x1000;
     pub const REGISTERS_SIZE: usize = 0x10;
     pub const STACK_SIZE: usize = 0x10;
 
@@ -20,7 +20,7 @@ impl Cpu {
     pub fn new() -> Cpu {
         Cpu {
             registers: [0; Cpu::REGISTERS_SIZE],
-            ram: [0; Cpu::RAM_SIZE],
+            ram: [0; Cpu::RAM_SIZE as usize],
             stack: [0; Cpu::STACK_SIZE],
             pc: 0x200,
             sp: 0,
